+++ src/main/java/com/ll/gooHaeYu/domain/payment/payment/dto/request/PaymentResDto.java
@@ -1,4 +1,4 @@
-package com.ll.gooHaeYu.domain.payment.payment.dto;
+package com.ll.gooHaeYu.domain.payment.payment.dto.request;
 
 import com.ll.gooHaeYu.domain.payment.payment.entity.type.PayType;
 import jakarta.validation.constraints.NotBlank;
